+++ pytorch_lightning/callbacks/xla_stats_monitor.py
@@ -76,7 +76,7 @@ class XLAStatsMonitor(Callback):
         if isinstance(trainer.accelerator, TPUAccelerator):
             raise MisconfigurationException(
                 "You are using XLAStatsMonitor but are not running on TPU."
-                f" The Trainer accelerator type is set to {trainer.accelerator.name().upper()}."
+                f" The accelerator type is set to {trainer.accelerator.name().upper()}."
             )
 
         device = trainer.strategy.root_device